SpringBooot Bank ATM Spring boot appplication

+++ 16bankappSpringBoot/BMS/src/main/java/com/example/dao/BankRepository.java
@@ -1,9 +1,15 @@
 package com.example.dao;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.model.Bank;
 
-public interface BankRepository extends JpaRepository<Bank, Long>{
+public interface BankRepository extends JpaRepository<Bank, BigDecimal>{
 
+	
+	
+	 //Bank findById(BigDecimal id);
+	
 }
